update A-MoreOOp
create Classes Storage, TaskList
