--- data/duke.txt
@@ -1,5 +1,0 @@
-T | 1 | read book
-D | 0 | return book | June 6th
-E | 1 | project meeting | Aug 6th 2-4pm
-T | 1 | join sports club
-T | 0 | like
