+++ src/main/java/duke/Duke.java
@@ -6,6 +6,7 @@ import duke.exceptions.InvalidKeywordException;
 import duke.tasks.Deadline;
 import duke.tasks.Event;
 import duke.tasks.Task;
+import duke.tasks.TaskList;
 import duke.tasks.ToDo;
 
 
@@ -18,28 +19,20 @@ import java.io.IOException;
 
 public class Duke {
 
-    private static ArrayList<Task> tasks = new ArrayList<>();
-    private static int taskCount = 0;
     private static String userInputLine;
     private static String command;  
     private static String description;
     private static String taskName;
     private static String taskTime;
     private static Ui ui;
+    private static TaskList tasks;
+    private static Storage storage;
 
     public static void main(String[] args) throws IOException {
         ui = new Ui();
-
-        //auto load
-        File autoSavedFileDir = new File("data");
-        File autoSavedFile = new File("data/duke.txt");
-        ui.showLoadingProcess();
-        if(checkFileExistence(autoSavedFileDir,autoSavedFile)){
-            ui.showLoadingFinished();
-        }
-        else {
-            ui.showNewFileDetected();
-        }
+        storage = new Storage();
+        tasks = new TaskList();
+        storage.init(tasks);
 
         Scanner in = new Scanner(System.in);
         boolean isQuit = false;
@@ -53,31 +46,31 @@ public class Duke {
 
             switch (command.toUpperCase()) {
             case "LIST":
-                listTasks();
+                tasks.listTasks();
                 break;
             case "DONE":
                 if(checkDescription() && checkTaskNum()) {
-                    doneTask(Integer.parseInt(description));
+                    tasks.doneTask(Integer.parseInt(description));
                 }
                 break;
             case "TODO":
                 if(checkDescription()) {
-                    addTask(new ToDo(description));
+                    tasks.addTask(new ToDo(description));
                 }
                 break;
             case "DEADLINE":
                 if(checkDescription() && checkDeadline()) {
-                    addTask(new Deadline(taskName,taskTime));
+                    tasks.addTask(new Deadline(taskName,taskTime));
                 }
                 break;
             case "EVENT":
                 if(checkDescription() && checkEvent()) {
-                    addTask(new Event(taskName, taskTime));
+                    tasks.addTask(new Event(taskName, taskTime));
                 }
                 break;
             case "DELETE":
                 if (checkDescription() && checkTaskNum()){
-                    deleteTask(Integer.parseInt(description));
+                    tasks.deleteTask(Integer.parseInt(description));
                 }
                 break;
             case "BYE":
@@ -92,7 +85,7 @@ public class Duke {
                 break;
             }
         }
-        autoSave(autoSavedFile);
+        storage.autoSave(storage.autoSavedFile,tasks);
     }
 
     private static void processInput(Scanner in) {
@@ -100,30 +93,12 @@ public class Duke {
             command = processCommand();
     }
 
-    public static void addTask(Task t) {
-        tasks.add(t);
-        taskCount = tasks.size();
-        ui.showAddTaskSuccessful(t.toString(), taskCount);
-    }
-
-    public static void addTaskSilent(Task t) {
-        tasks.add(t);
-        taskCount = tasks.size();
-    }
-
-    public static void listTasks() {
-        ui.sayShowTaskList();
-        for (int i = 0; i< taskCount; i++) {
-            System.out.println((i+1)+"."+tasks.get(i).toString());
-        }
-        ui.showSplitter();
-    }
 
     public static Boolean checkTaskNum() {
         boolean isNumValid = true;
         try{
             int inputNum = Integer.parseInt(description);
-            if (inputNum > taskCount || taskCount == 0 || inputNum == 0) {
+            if (inputNum > tasks.taskCount || tasks.taskCount == 0 || inputNum == 0) {
                 throw new DukeException();
             }
         }catch (NumberFormatException e) {
@@ -134,21 +109,6 @@ public class Duke {
             ui.printNotSetTaskNumError();
         }
         return isNumValid;
-    }
-    public static void doneTask(int taskIndex) {
-        tasks.get(taskIndex - 1).markAsDone();
-        ui.showTaskDone(tasks.get(taskIndex - 1).toString());
-    }
-
-    public static void doneTaskSilent(int taskIndex) {
-        tasks.get(taskIndex - 1).markAsDone();
-    }
-
-    public static void deleteTask(int taskIndex) {
-        ui.showTaskDeleting(tasks.get(taskIndex - 1).toString());
-        tasks.remove(taskIndex - 1);
-        taskCount = tasks.size();
-        ui.showTaskNum(taskCount);
     }
 
 
@@ -227,83 +187,6 @@ public class Duke {
         }
         else if (taskTime.isEmpty()) {
             throw new EmptyDescriptionException();
-        }
-    }
-
-    public static boolean checkFileExistence(File dirName,File fileName){
-        boolean isExist = false;
-        try{
-            autoLoad(fileName);
-            isExist = true;
-        }catch(FileNotFoundException e){
-            if(fileName.isDirectory()) {
-                ui.showMissingFile();
-            }
-            else{
-                ui.showMissingFolder();
-                createDirectory(dirName);
-            }
-            createSavedFile(fileName);
-        }
-        return isExist;
-    }
-    public static void autoLoad(File fileName) throws FileNotFoundException {
-        if(fileName.exists()){
-            Scanner rs= new Scanner(fileName);
-            while(rs.hasNext()){
-                String record = rs.nextLine();
-                processRecord(record);
-            }
-        }
-        else {
-            throw new FileNotFoundException();
-        }
-    }
-    public static void autoSave(File fileName) throws IOException{
-        String filePath = fileName.getPath();
-        new FileWriter(fileName, false).close();
-        FileWriter fw = new FileWriter(filePath,true);
-        for(int i = 0; i< tasks.size(); i++){
-            fw.write(tasks.get(i).saveAsText() + System.lineSeparator());
-        }
-        fw.close();
-    }
-
-    public static void createDirectory(File dir){
-        boolean isCreated = dir.mkdirs();
-        if(isCreated){
-            ui.showDirCreatedSucceed();
-        }
-        else{
-            ui.showDirCreatedFail();
-        }
-    }
-
-    public static void createSavedFile(File file){
-        try {
-            file.createNewFile();
-        }catch (IOException e){
-            ui.showFileCreatedFail();
-        }
-    }
-
-    public static void processRecord(String record){
-        String[] recordInfos = record.split(" \\| ");
-        switch (recordInfos[0]){
-        case "T":
-            addTaskSilent(new ToDo(recordInfos[2]));
-            break;
-        case "D":
-            addTaskSilent(new Deadline(recordInfos[2], recordInfos[3]));
-            break;
-        case "E":
-            addTaskSilent(new Event(recordInfos[2], recordInfos[3]));
-            break;
-        default:
-            break;
-        }
-        if(recordInfos[1].equals("1")) {
-            doneTaskSilent(tasks.size());
         }
     }
